Move pipeline to it's own module
